Clase060125/VisualizacionSQLIntro/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
slowapi==0.1.9
